Remove WasExternal::DoGlobal
There were two places in Yiqun where this flag was used, executing
code when DoGlobal was true:
1) Yiqun::Fit2PhotonClust() performs global fit to recalculate chi2
2) Yiqun::FitEvent() performs global fit when nRealClusts > 1
DoGlobal was always set as true in the code up to this point.
If DoGlobal were set to false neither of the above things would
happen; in particuler there would be no fitting when nRealClusts > 2.
As the default (true) seems like the sensible option, we remove the
flag entirely.