Remove error-related factors from WasExternal
Remove errFactor, errQ, Power1, Power2 and UseThis_Err members. This
change does introduce a slight difference in behaviour to the prior
state of the code. This is because previously the UseThis_Err flag
was always false. This caused a particular error calculation to be used
in FitTower::Fcn1(). However, looking at Steve's code his macros
typically call FitTower::Setwe_ErrFactors() to set the first 4 values.
This causes UseThis_Err to actually always be *true*. The result is a
slightly different error calculation being used. Because we still use
the default values for errFactor etc the difference ends up being small
but it is there, so I note it here. However I think this change makes
the behaviour more closely resemble that of Steve's original package.
I also deleted some other related code, such as parts that are no longer
executed, including the previously used error calculation.