Remove WasExternal::UseThis_ab
This flag was effectively always true. Therefore remove it and related
code. Set the defaults for WasExternal::a and ::b parameters to the
appropriate values so they no longer need to be set in StFmsHitMaker.
The a and b parameters should be moved from WasExternal to FitTower at
some point, but I will do that later.