Remove WasExternal::Force2Mass
Steve occasionally in his original code used this, presumably to force
the photon-pair mass to the pi0 mass. However, he generally didn't.
That is the behaviour currently in the code and, I imagine, what we
want, so I removed the flag entirely. I eliminated associated code, as
it was never called.