Replace use of WasRandom::myRand
Yiqun used this random number generator in one place. It was a TRandom
object, which the ROOT developers warn is a bad random number
generator. It also had the same seed for every run. Therefore we remove
it, and use the global ROOT random generator via gRandom, which is a
superior TRandom3 generator.