Remove 'step' from WasExternal
This array was set in WasExternal then copied to FitTower without
modification. It is therefore now just stored in FitTower. While
used by Yiqun also, it can always be accessed via the FitTower that
Yiqun stores, so it doesn't need to be external. Yiqun did have its own
step array as well, but it didn't seem to be used for anything, so it
was also deleted.