Move EDepCorrection from WasExternal
This energy dependence function is now stored in Yiqun, where it is
primarily used. It is also accessible via a static function, so it can
still be called by HitCluster. However it still isn't clear to me
whether the HitCluster code using it is needed, so it may be removed
at a later date.