Moving energy matrices from hit maker
These should be placed in the point maker code, or even Yiqun itself,
as they are specific to that. For now I am placing them in the point
maker, but we could move them to Yiqun at a later date if it becomes
clear that is preferrable.

We also need to change access to StFmsDbMaker in our code from using the
global gStFmsDbMaker pointer to using GetMaker() (this came up in a
recent FMS software review), which requires accessing the make via its
name. Therefore I made this change, and I changed the StFmsDbMaker name
in the test macro to "fmsDb" for consistency with the STAR BFC chain.