Use trigger data to fill energy matrices
In my original version I read hits from the MuDST, but this turns out
to be inconsistent with Yuxi's implementation in StFmsHitMaker, where
he uses trigger data. I don't understand the difference - I thought the
MuDST should give the same - but for consistency in this debugging
phase I will keep it the same as Yuxi's version. We should think about
using the MuDST as the input for this, and try to figure out the reason
for the difference. Really, we probably need to support different types
of input, like the hit maker does - MuEvent, MuDST, trigger data etc,
to account for the different times the makes may be run (production,
or just someone regenerating hits with a new calibration).