Add utility function to TowerUtil
I want to see if I can simplify TowerUtil by reworking the cluster-
finding using built-in STL containers and algorithms rather than so
much hand-written code. Add a simple function to populate STL
containers from ROOT collections to aid in this, as I don't yet want to
change the interface to TowerUtil, and it works with TObjArray
arguments.