Redesign TowerUtil implementation
This tries to provide a briefer and more understandable implementation,
using STL containers and algorithms. It tries to break up the main
TowerUtil algorithm into smaller functions to make it easier to follow.

For example, the tower-cluster association occurs in a single function,
called within a loop, rather than being part of one long loop with other
loops inside it.