Remove use of TMatrix in Yiqun
The energy TMatrices were populated in StFmsPointMaker and passed to
Yiqun. However, they were then only used to populate a list of TowerFPD
objects in Yiqun, so it is simpler to just directly create those
TowerFPD lists in StFmsPointMaker and pass them to Yiqun, and avoid the
intermediate step of energy matrices.