Reduce amount of code in Geom
Use database structures directly, rather than copying data to other
arrays. This is briefer, and should be more easily maintainable. Also
delete methods and fields no longer used in this version of the project.

The interface has not changed, but it would make sense to stop using
east/west and NSTB ('north-south-top-bottom') to identify subdetectors
and instead use the detector ID in the database. This will require
changing files other than the Geom source files of course.