Rework tower-cluster assocation code
The logic of the algorithm is unchanged, and gives the same result.
However the implementation has changed a lot, using a class to store the
tower-cluster association information and make the necessary
calculations to do the association.

I think this is easier to follow, as each step of the association is now
in a separate, commented class method; and the code in the main function
is now shorted and easier to follow in the context of the rest of the
algorithm.