Begin to reduce length of Yiqun event fitting
This happens in a very long for loop that is hard to follow. Begin by
reducing the amount of whitespace in parts. One- and two-photon fitting
function now return chi2/ndf rather than chi2, as chi2/ndf is all that
is ever used. This also allows shortening of parts of the code in the
loop.