Add StFmsHit member to TowerFPD
Prototype this by adding an StFmsHit and removing just the energy
member of TowerFPD. Access to tower energy is now done via this hit. The
hit pointer is const to prevent changing it, which shouldn't be
necessary. Doing so required removing some HitCluster methods that
modify the hit energy, but these are never actually called in our
version of the code, so removing them doesn't affect behaviour, and
modifying the hit energy really shouldn't be necessary as far as I can
tell.