Add TowerFPD initialization method to set row/col
Place TowerFPD initialization into the class itself, as this keeps
things more encapsulated, and makes it easier to test for towers that
encounter a problem during initialization and consequently should be
skipped for clustering.

This required a reworking for the cluster tower search in Yiqun as the
TowerFPD no longer takes a row/column argument in its constructor.