Fix cluster fit parameter upper limits
The 2-photon and global fits both had errors for the upper limit of the
number of photons in the fit - they set the lower limit to the upper
limit, and left the upper limit unset. Fixing this doesn't actually
change the results, so I infer from that the fit parameters limits
aren't really used.