Inherit HitCluster from StFmsCluster
These two classes duplicate a lot of information, so use the simpler
StFmsCluster as the base class for HitCluster. Remove HitCluster members
that duplicate those from StFmsCluster, and use StFmsCluster methods to
access those members throughout the code. This also means we don't need
to implement methods for those members in both classes.