Discard events failing clustering
Rename Yiqun::Y() to Yiqun::cluster() and no longer call it in the
constructor. Make it now return bool (indicating success or error in
clustering) and call this function in StFmsPointMaker. This way we can
exclude events failing at the clustering stage.

Note that this leads to a change in QA plot output from previously, as
we now exclude a small number of events (~3% level).

Resolves #5.