Omit clusters reconstructed at x, y = 0, 0
The (0, 0) coordinates of these clusters is invariant under the change
in row counting convention (top-down to bottom-up). This leads me to
believe that they are bad clusters, or are at least not suitable for
analysis. The (0, 0) coordinates seems to derive from the cluster
moment calculation, but I don't understand enough of what is going on
there to know exactly what the problem is.