Delete StFmsPointCollection
Use a simple vector of StFmsPoint pointers for the photon array in
StFmsCluster.

Note that it is not currently possible to use the STAR-standard
structures like StSPtrVecFmsPoint because of problems with including
StEvent/StContainers.h that I've not been able to resolve.

See https://www.star.bnl.gov/rt3/Ticket/Display.html?id=2816