Remove row flip before channel validity test
This was a holdover from an earlier version, where we passed row number
to Steve's code, where row number was defined the opposite way round to
how it is in the database (top-down instead of bottom-up). As that is no
longer the case we can remove it, because the detector is vertically
symmetric and the validity test therefore is insensitive to the flip.