Reset TProcessID object count after each event
This is done for now in StFmsPointMaker. It should probably be done
somewhere else, e.g. in StMuDstMaker, but I need to confirm with the
software team where would be best, so until the software review it can
stay here. It shouldn't interfere with anything else as I don't think
any other STAR makers use TRef.