Set StFmsPointMaker ClassDefs to 0
The old ClassDef values reflected the development history of the PSU
package. As this code is new, they should be reset.

None of the classes in this package are designed to be written to file.
Indeed, most of them don't possess any mechanism to persistently write
their data in a meaningful way. Therefore we set them all to zero to
prevent object I/O.

Closes #25.