Replace Boost ptr_list with list of unique_ptr
Clustering was previously done using a Boost ptr_list (from the
ptr_container library). As we are now migrating to C++11, replace this
with std::list<std::unique_ptr>>. Note that we no longer need the space
between right angle brackets in the template declaration thanks to
C++11!

This is the initial step in phasing out Boost in favour of C++11 as
discussed in #34.