Remove BOOST_FOREACH
This is the other task in removing Boost, as discussed in #34.

Unfortunately, as we are currently developing for gcc 4.4.7, range-based
for loops are not yet supported. Therefore we are limited to traditional
for loops using iterators. At least the use of auto types helps reduce
some of the associated verbosity in some places.

When we move to gcc 4.6 or later (where range-based loops are supported)
it might be nice to clean up the modified code by using them.

Closes #34.