Replaces auto_ptr with unique_ptr
auto_ptr is deprecated in C++11, so replaces with unique_ptr in
accordance with #34.

This causes a complication with CINT dictionary generation. Because
unique_ptr is non-copyable it makes StFmsTowerCluster non-copyable. When
a CINT dictionary is generated this yields a "deleted function" error
when generating a default copy constructor.

For simplicity we simply disallow copying of the class, and declare a
do-nothing private copy constructor and assignment operator.

Closes #34.