Replace NULL with nullptr
gcc 4.4, for which we are currently developing, does not actually
support nullptr. However, we can use it just as a substitute for NULL as
ROOT defines nullptr (just as 0) in Rtypes.h.

Once we move to gcc 4.6+ nullptr proper will be supported by the
compiler, and the current implementation should still be supported.