Removes check on number of photons in global refit
This check is also made in the globalFit function so it is not necessary
to repeat it.

This does change the logic a bit, as an event with more than the maximum
allowed number of photons for a global fit will now be accepted. This
makes sense to me as the inability to do a global fit doesn't invalidate
the original per-cluster fit.