Changes 2-photon fit arguments to arrays
This enforces the fixed number of parameters (7), which isn't the case
with vector. It is also more self-documenting.

As a result the setMinuitParameter() method is made a template method so
the same code can be used with vector and array arguments (for n- and
2-photon fits respectively).

std::array has to be hidden from CINT, as CINT can't deal with C++11.